Create example/test/, move test for project_initSpi() to it
