+++ example/firmware/init_spi.c
