+++ example/test/testExample.cpp
@@ -1,0 +1,5 @@
+#include <CppUTest/CommandLineTestRunner.h>
+
+int main(int argCount, char* argv[]) {
+  return RUN_ALL_TESTS(argCount, argv);
+}
