+++ run-tests.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 cmake -B test/build -S test && \
-  cmake --build test/build --target testFirment && test/build/testFirment
+  cmake --build test/build --target testFirment --target testExample && \
+  echo "Running testFirment" && test/build/testFirment && \
+  echo "Running testExample" && test/build/example-test/testExample
